reorganise /benchmarks
