+++ benchmarks/src/echo/echo-bench.ts
